requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
tensorflow==2.17.1
scikit-learn==1.4.2
matplotlib==3.8.2
